feat/#19 UserRoom에 이름이 잘못 명시되어 있던것을 수정

+++ src/main/java/woosap/Pepple/repository/UserRoomRepository.java
@@ -7,5 +7,5 @@ import woosap.Pepple.entity.UserRoom;
 
 public interface UserRoomRepository extends JpaRepository<UserRoom, String> {
 
-    long countByRoomId(long roomId);
+    int countByRoomId(long roomId);
 }
